refactor: Relocate authentication states and models to account-manager API
* Refactor the authentication logic by moving core auth models and states from `traktauth` to a more generic `accountmanager` module to support multi-provider authentication.

    * refactor: Move `AuthState`, `AuthError`, and `TokenRefreshResult` from `traktauth.api` to `accountmanager.api`
    * refactor: Rename `TraktAuthState` to `AccountAuthState` and relocate it to the `accountmanager` module
    * refactor: Update `UserProfile` and `ObserveUserProfileInteractor` to use the new `AccountAuthState`
    * chore: Update `TraktHttpClient` to reference the relocated `TokenRefreshResult`
    * chore: Update dependencies in `features/debug/presenter` and `ios-framework` to include `account-manager-api`
    * refactor: Update presenters (`DebugPresenter`, `ProfilePresenter`, `RootPresenter`) and their corresponding tests to use the new auth state and model locations
    * refactor: Update test stubs and integration scenarios to reflect the `TraktAuthState` to `AccountAuthState` renaming
