+++ domain/episode/src/commonTest/kotlin/com/thomaskioko/tvmaniac/domain/episode/PendingUploadsWorkerTest.kt
@@ -1,13 +1,13 @@
 package com.thomaskioko.tvmaniac.domain.episode
 
 import app.cash.turbine.test
+import com.thomaskioko.tvmaniac.accountmanager.api.AccountAuthState
 import com.thomaskioko.tvmaniac.core.logger.fixture.FakeLogger
 import com.thomaskioko.tvmaniac.core.tasks.api.WorkerResult
 import com.thomaskioko.tvmaniac.data.library.testing.FakeLibraryRepository
 import com.thomaskioko.tvmaniac.episodes.testing.FakeWatchedEpisodeSyncRepository
 import com.thomaskioko.tvmaniac.syncstate.api.SyncError
 import com.thomaskioko.tvmaniac.syncstate.testing.FakeSyncObserver
-import com.thomaskioko.tvmaniac.traktauth.api.TraktAuthState
 import com.thomaskioko.tvmaniac.traktauth.testing.FakeTraktAuthRepository
 import io.kotest.matchers.shouldBe
 import io.kotest.matchers.types.shouldBeInstanceOf
@@ -32,7 +32,7 @@ internal class PendingUploadsWorkerTest {
 
     @Test
     fun `should return Success when user is logged in and sync completes`() = runTest {
-        authRepository.setState(TraktAuthState.LOGGED_IN)
+        authRepository.setState(AccountAuthState.LOGGED_IN)
         syncRepository.setPendingEpisodesError(null)
 
         val result = worker.doWork()
@@ -42,7 +42,7 @@ internal class PendingUploadsWorkerTest {
 
     @Test
     fun `should return Success without syncing when user is logged out`() = runTest {
-        authRepository.setState(TraktAuthState.LOGGED_OUT)
+        authRepository.setState(AccountAuthState.LOGGED_OUT)
         syncRepository.setPendingEpisodesError(RuntimeException("should never run"))
 
         val result = worker.doWork()
@@ -53,7 +53,7 @@ internal class PendingUploadsWorkerTest {
 
     @Test
     fun `should flush pending followed shows when user is logged in`() = runTest {
-        authRepository.setState(TraktAuthState.LOGGED_IN)
+        authRepository.setState(AccountAuthState.LOGGED_IN)
 
         worker.doWork().shouldBeInstanceOf<WorkerResult.Success>()
 
@@ -62,7 +62,7 @@ internal class PendingUploadsWorkerTest {
 
     @Test
     fun `should return Retry when syncPendingEpisodes throws`() = runTest {
-        authRepository.setState(TraktAuthState.LOGGED_IN)
+        authRepository.setState(AccountAuthState.LOGGED_IN)
         syncRepository.setPendingEpisodesError(RuntimeException("network down"))
 
         val result = worker.doWork()
@@ -73,7 +73,7 @@ internal class PendingUploadsWorkerTest {
 
     @Test
     fun `should retry once and succeed when error cleared between attempts`() = runTest {
-        authRepository.setState(TraktAuthState.LOGGED_IN)
+        authRepository.setState(AccountAuthState.LOGGED_IN)
 
         syncRepository.setPendingEpisodesError(RuntimeException("flaky"))
         worker.doWork().shouldBeInstanceOf<WorkerResult.Retry>()
@@ -84,7 +84,7 @@ internal class PendingUploadsWorkerTest {
 
     @Test
     fun `should log BackgroundSyncFailed given syncPendingEpisodes throws`() = runTest {
-        authRepository.setState(TraktAuthState.LOGGED_IN)
+        authRepository.setState(AccountAuthState.LOGGED_IN)
         val cause = RuntimeException("rate limit 429")
         syncRepository.setPendingEpisodesError(cause)
