+++ ios-framework/build.gradle.kts
@@ -33,6 +33,7 @@ scaffold {
             export(projects.core.featureFlags.api)
             export(projects.navigation.api)
             export(projects.domain.theme)
+            export(projects.data.accountManager.api)
             export(projects.data.traktauth.api)
             export(projects.api.trakt.api)
             export(projects.features.calendar.presenter)
@@ -92,6 +93,7 @@ kotlin {
                 api(projects.core.networkUtil.implementation)
                 api(projects.i18n.api)
                 api(projects.navigation.api)
+                api(projects.data.accountManager.api)
                 api(projects.data.traktauth.api)
                 api(projects.features.calendar.presenter)
                 api(projects.features.discover.presenter)
